pyproject.toml
[tool.poetry]
name = "scrapingdatasensorproject"
version = "0.1.0"
description = ""
authors = ["MossOcelot <phuminforwork@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.0.0"
pillow = "^10.4.0"
opencv-python = "^4.10.0.84"
pyqt5-tools = "^5.15.9.3.3"
pyautogui = "^0.9.54"
pyinstaller = { version = "^6.8.0", python = ">=3.8,<3.13" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
